Move some lints from `Misc` to `Operators`
